add request per second metric

+++ Transformer/Library/src/main/java/mobilecloud/metric/Metric.java
@@ -8,9 +8,9 @@ public class Metric implements Serializable {
     public final int readBPS;
     public final int writeBPS;
     public final double cpuLoadPercentage;
-    public final int requestPerSecond;
+    public final double requestPerSecond;
     
-    public Metric(int readBPS, int writeBPS, double cpuLoadPercentage, int requestPerSecond) {
+    public Metric(int readBPS, int writeBPS, double cpuLoadPercentage, double requestPerSecond) {
         this.readBPS = readBPS;
         this.writeBPS = writeBPS;
         this.cpuLoadPercentage = cpuLoadPercentage;
@@ -19,7 +19,8 @@ public class Metric implements Serializable {
 
     @Override
     public String toString() {
-        return "readBPS: " + readBPS + "bps, writeBPS: " + writeBPS + "bps, cpuLoad: " + cpuLoadPercentage + "%, requestPerSecond: " + requestPerSecond;
+        return "readBPS: " + readBPS + "bps, writeBPS: " + writeBPS + "bps, cpuLoad: " + cpuLoadPercentage
+                + "%, requestPerSecond: " + requestPerSecond;
     }
     
 }
